+++ src/main/java/com/my/project/sign/service/SignUpService.java
@@ -3,5 +3,5 @@ package com.my.project.sign.service;
 import com.my.project.sign.dto.SignUpRequestDto;
 
 public interface SignUpService {
-    int signUp(SignUpRequestDto SignUpRequestDto);
+    Long signUp(SignUpRequestDto SignUpRequestDto);
 }
